Rename getAllRecords Lambda Function
